crud final schedule management

+++ app/Http/Controllers/Admin/StudioScheduleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\ScheduleModel;
+use App\Models\SubScheduleModel;
 use App\Http\Requests\Admin\StudioSchedulesRequest;
 use Illuminate\Http\Request;
 use Alert;
@@ -17,7 +18,10 @@ class StudioScheduleController extends Controller
      */
     public function index()
     {
-        $schedules = ScheduleModel::all();
+        $schedules = ScheduleModel::with('subSchedule')
+                                ->orderBy('date', 'desc')
+                                ->get();
+                            
         return view('pages.admin.StudioSchedule.index', ['schedules' => $schedules]);
     }
 
@@ -39,15 +43,43 @@ class StudioScheduleController extends Controller
      */
     public function store(StudioSchedulesRequest $request)
     {
+        $dateNow = now()->format('m/d/Y');
+        $datePicked = $request->date;
+        $date = date_create($datePicked);
+        $dateConverted = date_format($date, 'Y-m-d');
+
+        $openHour = $request->open_hours;
+        $closeHour = $request->close_hour;
+        $startHour = date('H', strtotime($openHour));
+        $finishHour = date('H', strtotime($closeHour));
+        $workingHours = $finishHour - $startHour;
+        $minWorkingHours = 5;
 
         if($request->validated()):
-            $date = date_create($request->date);
-            $dateConverted = date_format($date, 'Y-m-d');
+            if($datePicked < $dateNow):
+                Alert::toast('Date cannot less than date now', 'error');
+                return redirect()->back();
+            endif;
+
+            if(ScheduleModel::where('date', $dateConverted)->exists()):
+                Alert::toast('Date already exist, please choose another date', 'error');
+                return redirect()->back();
+            endif;
+
+            if($closeHour < $openHour):
+                Alert::toast('Close hour cannot less than open hour', 'error');
+                return redirect()->back();
+            endif;
+
+            if($workingHours <= $minWorkingHours ):
+                Alert::toast('Operational hour cannot less than 5 hours', 'error');
+                return redirect()->back();
+            endif;
+           
             ScheduleModel::create([
                 'date' => $dateConverted,
-                'session_available' => $request->session_available,
-                'session_reserved' => $request->session_reserved,
-                'open_hours' => $request->open_hours
+                'open_hours' => $openHour,
+                'close_hour' => $closeHour
             ]);
             
         else:
@@ -93,9 +125,49 @@ class StudioScheduleController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(StudioSchedulesRequest $request, $id)
     {
-        //
+        $dateNow = now()->format('m/d/Y');
+        $datePicked = $request->date;
+        $date = date_create($datePicked);
+        $dateConverted = date_format($date, 'Y-m-d');
+
+        $openHour = $request->open_hours;
+        $closeHour = $request->close_hour;
+
+        $reservedScheduleInSameDateExist = SubScheduleModel::with('schedule')
+                                                    ->where('schedule_id', $id)
+                                                    ->select('studio_reserved')    
+                                                    ->exists();
+        if($request->validated()):
+            if($datePicked < $dateNow):
+                Alert::toast('Date cannot less than date now', 'error');
+                return redirect()->back();
+            endif;
+
+            if($reservedScheduleInSameDateExist):
+                Alert::toast('Data cannot be changed, some schedules already reserved', 'error');
+                return redirect()->back();
+            endif;
+
+            if($closeHour < $openHour):
+                Alert::toast('Close hour cannot less than open hour', 'error');
+                return redirect()->back();
+            endif;
+
+            $scheduleData = ScheduleModel::findOrFail($id);
+            $scheduleData->update([
+                'date' => $dateConverted,
+                'open_hours' => $openHour,
+                'close_hour' => $closeHour
+            ]);
+
+            Alert::toast('Data berhasil diubah', 'success');
+            return redirect()->route('schedules.index');
+        endif; 
+
+        Alert::toast('Gagal mengubah data', 'error');
+        return redirect()->route('schedules.index');
     }
 
     /**
@@ -107,6 +179,15 @@ class StudioScheduleController extends Controller
     public function destroy($id)
     {
         $schedule = ScheduleModel::findOrFail($id);
+        $reservedScheduleRunningExist = SubScheduleModel::with('schedule')
+                                            ->where('schedule_id', $id)
+                                            ->select('studio_reserved')    
+                                            ->exists();
+        if($reservedScheduleRunningExist):
+            Alert::toast('Data cannot be deleted, you have some schedules reserved running', 'error');
+            return redirect()->back();
+        endif;
+
         $schedule->delete();
 
         Alert::toast('Data berhasil dihapus', 'success');
